Modified main deploy script to be run using PHP CLI
The reason for this was that I had did not want to make my deployment script
accessible through a URI.

Therefore, my git hook simply ssh's into my deployment server and executes
the deploy.php script via the PHP CLI

In order for this to work, I had to detect if the access code were coming in
as program arguments.

I also added a site argument so that I only needed a single deploy.php
script and can have a seperate deploy-config script for each site/env pairing I have

There is always a better way, but this one is working for me.